+++ src/Smartstore.Core/Checkout/Tax/Services/TaxService.cs
@@ -7,11 +7,12 @@ using Smartstore.Core.Common.Services;
 using Smartstore.Core.Data;
 using Smartstore.Core.Identity;
 using Smartstore.Core.Localization;
+using Smartstore.Data.Hooks;
 using Smartstore.Engine.Modularity;
 
 namespace Smartstore.Core.Checkout.Tax
 {
-    public partial class TaxService : ITaxService
+    public partial class TaxService : AsyncDbSaveHook<TaxCategory>, ITaxService
     {
         const string DefaultTaxFormat = "{0} *";
 
@@ -42,6 +43,35 @@ namespace Smartstore.Core.Checkout.Tax
             _localizationService = localizationService;
             _taxSettings = taxSettings;
         }
+
+        #region Hook 
+
+        protected override Task<HookResult> OnDeletedAsync(TaxCategory entity, IHookedEntity entry, CancellationToken cancelToken)
+            => Task.FromResult(HookResult.Ok);
+
+        public override async Task OnAfterSaveCompletedAsync(IEnumerable<IHookedEntity> entries, CancellationToken cancelToken)
+        {
+            var deletedTaxCategoryIds = entries
+                .Where(x => x.InitialState == EntityState.Deleted)
+                .Select(x => x.Entity)
+                .OfType<TaxCategory>()
+                .Select(x => x.Id)
+                .ToList();
+
+            if (deletedTaxCategoryIds.Count > 0)
+            {
+                var newTaxCategoryId = await _db.TaxCategories
+                    .OrderBy(x => x.DisplayOrder)
+                    .Select(x => x.Id)
+                    .FirstOrDefaultAsync(cancelToken);
+
+                await _db.Products
+                    .Where(x => deletedTaxCategoryIds.Contains(x.TaxCategoryId))
+                    .ExecuteUpdateAsync(x => x.SetProperty(p => p.TaxCategoryId, p => newTaxCategoryId), cancelToken);
+            }
+        }
+
+        #endregion
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public virtual Provider<ITaxProvider> LoadActiveTaxProvider()