+++ changelog.md
@@ -10,8 +10,8 @@
 ### Improvements
 
 - #580 Added caching for live currency exchange rates.
-
-- #767 Handle tier prices in depending prices module 
+- #767 Handle tier prices in depending prices module.
+- #378 Remove the assignments of products to a tax category when the tax category is deleted.
 
 ### Bugfixes